feat: deploy n-markets
